Remove Emacs -*- lines from C files
These lines tell Emacs which settings to use for the file.

"Mode: C" is obsolete: GNU Emacs has auto-detected .c and .h files as
C code for decades.

"Tab-Width: 4" was applied inconsistently to some files but not
others. It's not a good idea to change the tab width anyway, as it's
confusing and defeats the whole purpose of the tab character (it's
supposed to be a variable-width character). I will reformat the files
to use 4-space indents. This will preserve the visual appearance but
cause less headache.

"Lowercase: Yes" was applied only to some files. To my knowledge it
has no effect in modern GNU Emacs.

I found the files via:

    git ls-files | grep '\.[ch]$' | xargs grep -F -- '-*-'