Change pointer coding style
The WebKit coding style does pointers like this: 'char* foo'.  That's
very weird for C programmers since it misleads about the syntax of the
C language. In the syntax, the asterisk is actually part of the
variable name, not the type name. For example, this is legal:
'char *foo, **bar'.

The clang-format setting is "PointerAlignment: Right".